refactor: move BaseError, ErrorResponse to data/model, move ResultWrapper to domain/model
