+++ androidApp/src/main/java/com/example/tasky/android/login/viewmodel/RegisterViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.tasky.android.login.screen.RegisterScreenEvent
 import com.example.tasky.android.login.screen.RegisterScreenState
-import com.example.tasky.common.model.onError
-import com.example.tasky.common.model.onSuccess
+import com.example.tasky.common.domain.model.onError
+import com.example.tasky.common.domain.model.onSuccess
 import com.example.tasky.login.data.model.RegisterBody
 import com.example.tasky.login.domain.ILoginRepository
 import com.example.tasky.login.domain.util.Validator
