+++ shared/src/commonMain/kotlin/com/example/tasky/agenda/data/AgendaDataSource.kt
@@ -15,9 +15,9 @@ import com.example.tasky.agenda.data.model.UpdateEventBody
 import com.example.tasky.agenda.data.model.UpdateReminderBody
 import com.example.tasky.agenda.data.model.UpdateTaskBody
 import com.example.tasky.common.data.manager.HttpManager
+import com.example.tasky.common.data.model.BaseError
 import com.example.tasky.common.data.util.safeCall
-import com.example.tasky.common.model.BaseError
-import com.example.tasky.common.model.ResultWrapper
+import com.example.tasky.common.domain.model.ResultWrapper
 import io.ktor.client.HttpClient
 import io.ktor.client.plugins.resources.delete
 import io.ktor.client.plugins.resources.get
