+++ shared/src/commonMain/kotlin/com/example/tasky/login/domain/LoginRepository.kt
@@ -1,7 +1,7 @@
 package com.example.tasky.login.domain
 
-import com.example.tasky.common.model.BaseError
-import com.example.tasky.common.model.ResultWrapper
+import com.example.tasky.common.data.model.BaseError
+import com.example.tasky.common.domain.model.ResultWrapper
 import com.example.tasky.login.data.LoginDataSource
 import com.example.tasky.login.data.model.LoginBody
 import com.example.tasky.login.data.model.LoginResponse
