+++ shared/src/commonMain/kotlin/com/example/tasky/login/domain/manager/LoginManager.kt
@@ -1,8 +1,8 @@
 package com.example.tasky.login.domain.manager
 
 import com.example.tasky.common.data.manager.HttpManager
-import com.example.tasky.common.model.ResultWrapper
-import com.example.tasky.common.model.onSuccess
+import com.example.tasky.common.domain.model.ResultWrapper
+import com.example.tasky.common.domain.model.onSuccess
 import com.example.tasky.dataStore.SettingsKey
 import com.example.tasky.login.data.model.LoginBody
 import com.example.tasky.login.domain.ILoginRepository
